+++ app/src/main/res/layout/activity_main.xml
@@ -7,72 +7,4 @@
         android:layout_height="match_parent"
         tools:context=".MainActivity">
 
-    <SearchView android:id="@+id/main_layout_searchView"
-                android:layout_width="0dp"
-                android:layout_height="wrap_content"
-                app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"
-                app:layout_constraintEnd_toStartOf="@+id/vertical_guideline" android:layout_marginEnd="8dp"
-                app:layout_constraintTop_toTopOf="parent"
-                app:layout_constraintHorizontal_bias="1.0"
-                app:layout_constraintBottom_toTopOf="@+id/horizontal_guideline1"/>
-
-    <ImageButton android:id="@+id/main_layout_menu_button"
-                 android:layout_width="0dp"
-                 android:layout_height="wrap_content"
-                 app:srcCompat="@drawable/main_layout_menu_button"
-                 android:scaleType="fitCenter"
-                 android:background="@android:color/transparent"
-                 app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-                 app:layout_constraintStart_toStartOf="@+id/vertical_guideline" android:layout_marginTop="8dp"
-                 app:layout_constraintTop_toTopOf="parent"
-                 android:layout_marginBottom="8dp" app:layout_constraintBottom_toTopOf="@+id/horizontal_guideline1"/>
-    <ImageView
-            android:id="@+id/dokuro_id1"
-            android:src="@drawable/dokuro_black"
-            android:layout_width="100dp"
-            android:layout_height="100dp"
-            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"
-            app:layout_constraintBottom_toTopOf="@+id/horizontal_guideline2"
-            android:layout_marginTop="8dp" app:layout_constraintTop_toTopOf="@+id/horizontal_guideline1"/>
-
-    <TextView
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:text="10%"
-            android:textSize="50sp"
-            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-            android:layout_marginBottom="8dp" app:layout_constraintBottom_toTopOf="@+id/horizontal_guideline2"
-            app:layout_constraintStart_toEndOf="@+id/dokuro_id1" android:layout_marginStart="8dp"
-            android:layout_marginTop="8dp" app:layout_constraintTop_toTopOf="@+id/horizontal_guideline1"/>
-
-
-    <android.support.v7.widget.RecyclerView
-            android:id="@+id/main_recyclerView"
-            android:layout_width="0dp"
-            android:layout_height="0dp"
-            app:layout_constraintTop_toTopOf="@+id/horizontal_guideline2"
-            android:layout_marginBottom="8dp" app:layout_constraintBottom_toBottomOf="parent"
-            app:layout_constraintEnd_toEndOf="parent" android:layout_marginEnd="8dp"
-            app:layout_constraintStart_toStartOf="parent" android:layout_marginStart="8dp"/>
-
-    <android.support.constraint.Guideline
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/vertical_guideline"
-            android:orientation="vertical"
-            app:layout_constraintGuide_percent="0.9"/>
-
-    <android.support.constraint.Guideline
-    android:layout_width="wrap_content"
-    android:layout_height="wrap_content"
-    app:layout_constraintGuide_percent="0.1"
-    android:orientation="horizontal" android:id="@+id/horizontal_guideline1"/>
-
-    <android.support.constraint.Guideline
-            android:layout_width="wrap_content"
-            android:layout_height="wrap_content"
-            android:id="@+id/horizontal_guideline2"
-            android:orientation="horizontal"
-            app:layout_constraintGuide_percent="0.3"/>
-
 </android.support.constraint.ConstraintLayout>
